backend/requirements.txt
flask
flask-cors
openai
pillow
rembg
torchvision
torch
numpy
requests
